arm: dts: Remove MSM watchdog IPI ping in msm8996
This is a debugging patch for b/32371978 to test if
smp_call_function on idle CPUs can cause deadlock. It
can also give us a clue if a driver is misbehaving by
disabling interrupts for too long because then only that
driver's functionality will stop working instead of causing
a full device reboot.

Bug: 32371978
Change-Id: Iac665a16a58c72b8324e11a4685f29c2ce28ad3a
Signed-off-by: Harsh Shandilya <harsh@prjkt.io>
Signed-off-by: AyushR1 <ayushrathore12501@gmail.com>